+++ src/Console/Parameter/Argument.php
@@ -5,6 +5,9 @@ namespace Parable\Console\Parameter;
 class Argument extends Base
 {
     /** @var int|null */
+    protected $required;
+
+    /** @var int|null */
     protected $order;
 
     public function __construct(
@@ -15,6 +18,36 @@ class Argument extends Base
         $this->setName($name);
         $this->setRequired($required);
         $this->setDefaultValue($defaultValue);
+    }
+
+    /**
+     * Set whether this argument is required.
+     *
+     * @param int $required
+     *
+     * @return $this
+     * @throws \Parable\Console\Exception
+     */
+    public function setRequired($required)
+    {
+        if (!in_array(
+            $required,
+            [\Parable\Console\Parameter::PARAMETER_REQUIRED, \Parable\Console\Parameter::PARAMETER_OPTIONAL])
+        ) {
+            throw new \Parable\Console\Exception("Required must one of the PARAMETER_* constants.");
+        }
+        $this->required = $required;
+        return $this;
+    }
+
+    /**
+     * Return whether the parameter is required.
+     *
+     * @return bool
+     */
+    public function isRequired()
+    {
+        return $this->required === \Parable\Console\Parameter::PARAMETER_REQUIRED;
     }
 
     /**